Rename deleteThought -> deleteThoughtWithCursor
