+++ src/action-creators/index.ts
@@ -26,7 +26,7 @@ export const cursorUp = reducerToThunk<typeof reducers.cursorUp>('cursorUp')
 export const deleteAttribute = reducerToThunk<typeof reducers.deleteAttribute>('deleteAttribute')
 export const deleteData = reducerToThunk<typeof reducers.deleteData>('deleteData')
 export const deleteEmptyThought = reducerToThunk<typeof reducers.deleteEmptyThought>('deleteEmptyThought')
-export const deleteThought = reducerToThunk<typeof reducers.deleteThought>('deleteThought')
+export const deleteThoughtWithCursor = reducerToThunk<typeof reducers.deleteThoughtWithCursor>('deleteThoughtWithCursor')
 export const dragHold = reducerToThunk<typeof reducers.dragHold>('dragHold')
 export const editableRender = reducerToThunk<typeof reducers.editableRender>('editableRender')
 export const editing = reducerToThunk<typeof reducers.editing>('editing')
