+++ src/reducers/index.ts
@@ -19,7 +19,7 @@ export { default as cursorUp } from './cursorUp'
 export { default as deleteAttribute } from './deleteAttribute'
 export { default as deleteData } from './deleteData'
 export { default as deleteEmptyThought } from './deleteEmptyThought'
-export { default as deleteThought } from './deleteThought'
+export { default as deleteThoughtWithCursor } from './deleteThoughtWithCursor'
 export { default as dragHold } from './dragHold'
 export { default as dragInProgress } from './dragInProgress'
 export { default as editableRender } from './editableRender'
