+++ src/shortcuts/index.ts
@@ -12,7 +12,7 @@ export { default as cursorNext } from './cursorNext'
 export { default as cursorPrev } from './cursorPrev'
 export { default as cursorUp } from './cursorUp'
 export { default as deleteEmptyThoughtOrOutdent, deleteEmptyThoughtOrOutdentAlias } from './deleteEmptyThoughtOrOutdent'
-export { default as deleteThought, deleteAliases } from './delete'
+export { default as deleteThoughtWithCursor, deleteAliases } from './delete'
 export { default as exportContext } from './exportContext'
 export { default as extractThought } from './extractThought'
 export { default as home } from './home'
